includes the suggested scheduler
Includes the suggested scheduler (Ofelia), instead of leaving it as an
exercise for the reader. There are problems with having multiple
Ofelia's on a host. I would assume the idea is to have a single instance
running per host, but there are documented problems with that too.

https://github.com/mcuadros/ofelia/issues/84

Maybe something to be solved by docker-gen.

https://github.com/nginx-proxy/docker-gen